+++ CHANGELOG.md
@@ -8,7 +8,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 
 - problem: improve error message formatting and add HTTP status code to the message
-- retry: continue retrying for HTTP 5xx errors up to maximum retries instead of failing immediately
+- retry: continue retrying for HTTP 5xx errors up to three times instead of failing immediately
 
 ## [8.31.0]