+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [8.31.1]
+
 ### Changed
 
 - problem: improve error message formatting and add HTTP status code to the message
@@ -692,7 +694,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.31.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.31.1...HEAD
+[8.31.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.31.0...v8.31.1
 [8.31.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.30.0...v8.31.0
 [8.30.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.29.2...v8.30.0
 [8.29.2]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.29.1...v8.29.2